#2916 comment: moved ssm client to common.utilities to avoid circular imports
